Restructure the app to use a better layout so that everything isn't just flattened into the contract.rs file
