+++ src/core/state.rs
