+++ src/instantiate.rs
@@ -1,0 +1,1 @@
+pub mod init_contract;
