+++ src/lib.rs
@@ -1,9 +1,12 @@
 #![warn(clippy::all)]
+// Public modules
 pub mod contract;
-pub mod error;
-pub mod helper;
-pub mod make_payment;
-pub mod msg;
-pub mod oracle_approval;
-pub mod register_payable;
-pub mod state;
+pub mod core;
+pub mod execute;
+pub mod instantiate;
+pub mod query;
+pub mod util;
+
+// Conditional modules
+#[cfg(feature = "enable-test-utils")]
+pub mod test_utilities;
