--- src/make_payment.rs
@@ -1,3 +1,0 @@
-pub struct MakePaymentV1 {
-    pub payable_uuid: String,
-}
