--- src/oracle_approval.rs
@@ -1,7 +1,0 @@
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct OracleApprovalV1 {
-    pub payable_uuid: String,
-}
