+++ src/util/conversions.rs
@@ -1,0 +1,6 @@
+use crate::util::constants::ONE_HUNDRED;
+use cosmwasm_std::Decimal;
+
+pub fn to_percent(dec: Decimal) -> String {
+    (dec * ONE_HUNDRED).to_string()
+}
